+++ src/pkg.ts
@@ -1,52 +1,14 @@
 import { PackageJson } from "read-pkg";
-import { PkgJsonPossibleTypes } from "./json-types";
+import { BundlibOptions, BundlibPkgFlagOptions } from "./bundlib-options";
 import { BrowserBuildFormat } from "./types";
 
 export type PkgJsonRelevantOutputFields = "main" | "module" | "browser" | "types" | "typings";
 export type PkgJsonModuleOutputFields = Extract<PkgJsonRelevantOutputFields, "main" | "module" | "browser">;
 export type PkgJsonTypesOutputFields = Extract<PkgJsonRelevantOutputFields, "types">;
 
-export interface BundlibPkgInputOptions {
-  input: PkgJsonPossibleTypes;
-}
-
-export interface BundlibPkgFlagOptions {
-  sourcemap: PkgJsonPossibleTypes;
-  esModule: PkgJsonPossibleTypes;
-  interop: PkgJsonPossibleTypes;
-  extend: PkgJsonPossibleTypes;
-  equals: PkgJsonPossibleTypes;
-}
-
-export type ValidMinOption = PkgJsonModuleOutputFields | PkgJsonModuleOutputFields[] | boolean | null | undefined;
-
-export interface BundlibPkgOtherOptions {
-  browser: PkgJsonPossibleTypes;
-  name: PkgJsonPossibleTypes;
-  id: PkgJsonPossibleTypes;
-  globals: PkgJsonPossibleTypes;
-  sourcemap: PkgJsonPossibleTypes;
-  min: PkgJsonPossibleTypes;
-}
-
-export interface BundlibPkgRemovedOptions {
-  iife: PkgJsonPossibleTypes;
-  amd: PkgJsonPossibleTypes;
-  umd: PkgJsonPossibleTypes;
-}
-
-export type BundlibPkgOptions = Partial<
-  BundlibPkgInputOptions &
-  BundlibPkgFlagOptions &
-  BundlibPkgOtherOptions &
-  BundlibPkgRemovedOptions
->;
-
 export interface BundlibPkgJson extends PackageJson {
-  bundlib?: PkgJsonPossibleTypes;
+  bundlib?: BundlibOptions;
 }
-
-export type BundlibPkgFlagFields = keyof BundlibPkgFlagOptions;
 
 export type PkgJsonOutputFields = PkgJsonModuleOutputFields | PkgJsonTypesOutputFields;
 
@@ -66,13 +28,16 @@ export interface BrowserOptions {
   globals: Record<string, string> | null;
 }
 
+export type FlagOptions = Record<keyof BundlibPkgFlagOptions, boolean>;
+
 export interface AnalizedPkg {
   cwd: string;
   pkg: BundlibPkgJson;
   dependencies: BundlibDependencies;
   input: string;
   output: BundlibOutputFiles;
-  minify: MinifyOutOptions;
+  sourcemap: boolean | "inline";
   browser: BrowserOptions;
-  options: Record<BundlibPkgFlagFields, boolean>;
+  minify: MinifyOutOptions;
+  options: FlagOptions;
 }